pyproject.toml
[tool.poetry]
name = "person-app"
version = "0.1.0"
description = ""
authors = ["Thiago Beviláqua <fthiagobevilaqua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
django-extensions = "^3.2.3"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
prettyconf = "^2.2.1"

[tool.ruff]
line-length = 99
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ipdb = "^0.13.13"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
